+++ octgnFX/Octgn.Core/DataExtensionMethods/CardExtensionMethods.cs
@@ -39,7 +39,7 @@
 
             Uri uri = null;
 
-            var files = Directory.GetFiles(set.GetPackUri(), card.GetImageUri() + ".*").OrderBy(x=>x.Length).ToArray();
+            var files = Directory.GetFiles(set.GetImagePackUri(), card.GetImageUri() + ".*").OrderBy(x=>x.Length).ToArray();
             if (files.Length == 0) //Generate or grab proxy
             {
                 files = Directory.GetFiles(set.GetPackProxyUri(), card.GetImageUri() + ".png");