+++ octgnFX/Octgn/Controls/ContactList.xaml.cs
@@ -29,72 +29,58 @@ namespace Octgn.Controls
 
         public void RefreshList()
         {
-            User[] flist = Program.LobbyClient.Friends.OrderByDescending(x => x.Status == UserStatus.Online).ThenBy(x => x.UserName).ToArray();
-            Dispatcher.Invoke(new Action(() =>
-                                             {
-                                                 foreach (
-                                                     var f in stackPanel1.Children.OfType<FriendListItem>().ToArray())
-                                                 {
-                                                     f.MouseDoubleClick -= FMouseDoubleClick;
-                                                     f.Dispose();
-                                                 }
-                                                 foreach (
-                                                     var f in stackPanel1.Children.OfType<GroupChatListItem>().ToArray())
-                                                 {
-                                                     f.MouseDoubleClick -= GiMouseDoubleClick;
-                                                     f.Dispose();
-                                                 }
-                                                 stackPanel1.Children.Clear();
-                                                 foreach (FriendListItem f in flist.Select(u => new FriendListItem
-                                                                                                 {
-                                                                                                     ThisUser = u,
-                                                                                                     HorizontalAlignment
-                                                                                                         =
-                                                                                                         HorizontalAlignment
-                                                                                                         .
-                                                                                                         Stretch
-                                                                                                 }))
-                                                 {
-                                                     f.MouseDoubleClick += FMouseDoubleClick;
-                                                     stackPanel1.Children.Add(f);
-                                                 }
-                                                 foreach (var g in Program.LobbyClient.Chatting.Rooms.Where(x => x.IsGroupChat))
-                                                 {
-                                                     var gc = new GroupChatListItem(g);
-                                                     gc.MouseDoubleClick += GiMouseDoubleClick;
-                                                     stackPanel1.Children.Add(gc);
-                                                 }
-                                             }));
+            //User[] flist = Program.LobbyClient.Friends.OrderByDescending(x => x.Status == UserStatus.Online).ThenBy(x => x.UserName).ToArray();
+            //Dispatcher.Invoke(new Action(() =>
+            //                                 {
+            //                                     foreach (
+            //                                         var f in stackPanel1.Children.OfType<FriendListItem>().ToArray())
+            //                                     {
+            //                                         f.MouseDoubleClick -= FMouseDoubleClick;
+            //                                         f.Dispose();
+            //                                     }
+            //                                     foreach (
+            //                                         var f in stackPanel1.Children.OfType<GroupChatListItem>().ToArray())
+            //                                     {
+            //                                         f.MouseDoubleClick -= GiMouseDoubleClick;
+            //                                         f.Dispose();
+            //                                     }
+            //                                     stackPanel1.Children.Clear();
+            //                                     foreach (FriendListItem f in flist.Select(u => new FriendListItem
+            //                                                                                     {
+            //                                                                                         ThisUser = u,
+            //                                                                                         HorizontalAlignment
+            //                                                                                             =
+            //                                                                                             HorizontalAlignment
+            //                                                                                             .
+            //                                                                                             Stretch
+            //                                                                                     }))
+            //                                     {
+            //                                         f.MouseDoubleClick += FMouseDoubleClick;
+            //                                         stackPanel1.Children.Add(f);
+            //                                     }
+            //                                     foreach (var g in Program.LobbyClient.Chatting.Rooms.Where(x => x.IsGroupChat))
+            //                                     {
+            //                                         var gc = new GroupChatListItem(g);
+            //                                         gc.MouseDoubleClick += GiMouseDoubleClick;
+            //                                         stackPanel1.Children.Add(gc);
+            //                                     }
+            //                                 }));
         }
 
 
         private static void GiMouseDoubleClick(object sender, MouseButtonEventArgs e)
         {
-            var fi = sender as GroupChatListItem;
-            if (fi == null) return;
-            var room = Program.LobbyClient.Chatting.GetRoom(fi.ThisRoom.GroupUser, true);
-            //var cw = Program.ChatWindows.FirstOrDefault(x => x.Id == room.Rid);
-            //if(cw == null)
-            //{
-            //	cw = new Windows.ChatWindow(room);
-            //	Program.ChatWindows.Add(cw);
-            //}
-            //cw.Show();`
+            //var fi = sender as GroupChatListItem;
+            //if (fi == null) return;
+            //var room = Program.LobbyClient.Chatting.GetRoom(fi.ThisRoom.GroupUser, true);
         }
 
 
         private static void FMouseDoubleClick(object sender, MouseButtonEventArgs e)
         {
-            var fi = sender as FriendListItem;
-            if (fi == null) return;
-            var room = Program.LobbyClient.Chatting.GetRoom(fi.ThisUser);
-            //var cw = Program.ChatWindows.FirstOrDefault(x => x.Id == room.Rid);
-            //if(cw == null)
-            //{
-            //	cw = new Windows.ChatWindow(room);
-            //	Program.ChatWindows.Add(cw);
-            //}
-            //cw.Show();
+            //var fi = sender as FriendListItem;
+            //if (fi == null) return;
+            //var room = Program.LobbyClient.Chatting.GetRoom(fi.ThisUser);
         }
 
         #region Implementation of IDisposable