+++ octgnFX/Octgn/Controls/SubscribeMessage.xaml.cs
@@ -11,6 +11,7 @@ namespace Octgn.Controls
     using System.Windows;
     using System.Windows.Media.Animation;
 
+    using Octgn.Extentions;
     using Octgn.Library.Exceptions;
 
     using log4net;
@@ -22,15 +23,14 @@ namespace Octgn.Controls
     {
         internal static ILog Log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
 
-
-        public int TotalSeconds { get; set; }
-
         public SubscribeMessage()
         {
             Log.Info("Creating");
-            this.Opacity = 0d;
-            TotalSeconds = 30;
-            this.IsVisibleChanged += OnIsVisibleChanged;
+            if (!this.IsInDesignMode())
+            {
+                this.Opacity = 0d;
+                this.IsVisibleChanged += OnIsVisibleChanged;
+            }
 
             InitializeComponent();
             Log.Info("Created");
@@ -46,51 +46,10 @@ namespace Octgn.Controls
             if ((bool)dependencyPropertyChangedEventArgs.NewValue)
             {
                 Log.Info("Showing sub nag");
-                var da = new DoubleAnimation(0, 1, new Duration(TimeSpan.FromSeconds(.5)));
-                da.Completed += (o, args) => this.WaitStart();
+                var da = new DoubleAnimation(0, 1, new Duration(TimeSpan.FromSeconds(2)));
+                da.Completed += (o, args) => this.EnableButtons();
                 this.BeginAnimation(UIElement.OpacityProperty, da);
             }
-        }
-
-        private void WaitStart()
-        {
-            Log.Info("Wait started");
-            var task = new Task(WaitRun);
-            task.ContinueWith(WaitDone);
-            task.Start();
-        }
-
-        private void WaitRun()
-        {
-            Log.Info("Wait running");
-            var endtime = DateTime.Now.AddSeconds(TotalSeconds);
-            while (DateTime.Now < endtime)
-            {
-                var dif = new TimeSpan(endtime.Ticks - DateTime.Now.Ticks);
-                Dispatcher.Invoke(new Action(() => { this.ProgressBar.Value = dif.TotalSeconds; }));
-                Thread.Sleep(500);
-                Dispatcher.Invoke(new Action(
-                    () =>
-                        { if (!WindowManager.Main.IsActive) endtime = endtime.AddSeconds(.49); }))
-                ;
-            }
-        }
-
-        private void WaitDone(Task task)
-        {
-            Log.Info("Wait done");
-            if (task.IsFaulted) Log.Warn("wd", task.Exception);
-
-
-            Dispatcher.Invoke(
-                new Action(
-                    () =>
-                    {
-                        Log.Info("Hiding sub nag");
-                        var da = new DoubleAnimation(1, 0, new Duration(TimeSpan.FromSeconds(.5)));
-                        da.Completed += (o, args) => this.Visibility = Visibility.Collapsed;
-                        this.BeginAnimation(UIElement.OpacityProperty, da);
-                    }));
         }
 
         private void SubscribeClick(object sender, RoutedEventArgs e)
@@ -109,7 +68,27 @@ namespace Octgn.Controls
                                                                       MessageBoxButton.OK,
                                                                       MessageBoxImage.Asterisk);
                                                       }
+                                                      this.Close();
                                                   });
+        }
+
+        private void EnableButtons()
+        {
+            if (!Dispatcher.CheckAccess())
+            {
+                Dispatcher.Invoke(new Action(this.EnableButtons));
+                return;
+            }
+            ButtonGrid.IsEnabled = true;
+        }
+        private void DisableButtons()
+        {
+            if (!Dispatcher.CheckAccess())
+            {
+                Dispatcher.Invoke(new Action(this.DisableButtons));
+                return;
+            }
+            ButtonGrid.IsEnabled = false;
         }
 
         private void ShowSubscribeSite(SubType subtype)
@@ -120,6 +99,24 @@ namespace Octgn.Controls
             {
                 Program.LaunchUrl(url);
             }
+        }
+
+        private void CloseClick(object sender, RoutedEventArgs e)
+        {
+            this.Close();
+        }
+
+        private void Close()
+        {
+            if (!Dispatcher.CheckAccess())
+            {
+                Dispatcher.Invoke(new Action(this.Close));
+                return;
+            }
+            this.DisableButtons();
+            var da = new DoubleAnimation(1, 0, new Duration(TimeSpan.FromSeconds(2)));
+            da.Completed += (o, args) => this.Visibility = Visibility.Collapsed;
+            this.BeginAnimation(UIElement.OpacityProperty, da);
         }
     }
 }