+++ octgnFX/Skylabs.Lobby/Properties/AssemblyInfo.cs
@@ -36,5 +36,5 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 
-[assembly: AssemblyVersion("3.1.46.141")]
-[assembly: AssemblyFileVersion("3.1.46.141")]
+[assembly: AssemblyVersion("3.1.50.145")]
+[assembly: AssemblyFileVersion("3.1.50.145")]